Get data on market expectations for inflation over the next 12 months (Focus/BCB)

This function provides the extraction of data and statistics related to the expectations of economic indicators, specifically the market expectations for inflation in the next 12 months, made available by the Central Bank of Brazil's Expectations System through the Olinda API. The data comes from several financial institutions: banks, funds, research houses, etc. Important: arguments are case sensitive.

get_inflation_12m(
  indicator = NULL,
  first_date = Sys.Date() - 2 * 365,
  last_date = Sys.Date(),
  smoothed = NULL,
  be_quiet = FALSE,
  use_memoise = TRUE,
  do_parallel = FALSE
)

Arguments

indicator

A single string or a character vector with inflation indicators names, see details for possible values. Defaults to NULL.

first_date

A single string or Date in "YYYY-mm-dd" format. Default for 2 years prior to the current date.

last_date

A single string or Date in "YYYY-mm-dd" format. Default is the current date.

smoothed

A single string indicating if the statistics should be smoothed ("yes" or "no"). Defaults to NULL.

be_quiet

Logical. Should messages or warnings not be displayed? Defaults to FALSE.

use_memoise

Logical. Sets the use of caching system, creating a folder named "cache_bcb" by default. Default to TRUE.

do_parallel

Logical. For using parallel data importation. Defaults to FALSE.

Value

A tibble with data from BCB-Olinda Data Services.

Details

For periods for which there are no statistics, they will be omitted from the query.

Possible values for indicator argument: 'IGP-DI', 'IGP-M', 'INPC', 'IPA-DI', 'IPA-M', 'IPCA', 'IPCA-15', 'IPC-Fipe', 'IPCA Administrados', 'IPCA Alimentação no domicílio', 'IPCA Bens industrializados', 'IPCA Livres', 'IPCA Serviços'.

Author

Fernando da Silva <fernando@fortietwo.com>

Examples

df <- get_inflation_12m(
  indicator = "IPCA",
  first_date = Sys.Date() - 30,
  smoothed = "yes",
  use_memoise = FALSE
)
#> 
#> Fetching [IPCA] data from BCB-Olinda... 
#> 
#> Found 36 observations!